remove RepeatingManager
