+++ QSB/SectorSync/QSBSectorManager.cs
@@ -15,10 +15,27 @@ namespace QSB.SectorSync
 		public bool IsReady { get; private set; }
 		public List<QSBSector> FakeSectors = new();
 
-		private void OnEnable() => RepeatingManager.Repeatings.Add(this);
-		private void OnDisable() => RepeatingManager.Repeatings.Remove(this);
-
 		public List<BaseSectoredSync> SectoredSyncs = new();
+
+		#region repeating timer
+
+		private const float TimeInterval = 0.4f;
+		private float _checkTimer = TimeInterval;
+
+		private void Update()
+		{
+			_checkTimer += Time.unscaledDeltaTime;
+			if (_checkTimer < TimeInterval)
+			{
+				return;
+			}
+
+			Invoke();
+
+			_checkTimer = 0;
+		}
+
+		#endregion
 
 		public void Invoke()
 		{
